+++ client/src/main/java/net/oneandone/stool/client/Servers.java
@@ -94,13 +94,26 @@ public class Servers {
     }
 
     public List<Reference> list(String filter) throws IOException {
+        int idx;
         List<Reference> result;
         Client client;
+        String clientFilter;
+        String serverFilter;
 
+        idx = filter.lastIndexOf('@');
+        if (idx == -1) {
+            clientFilter = filter;
+            serverFilter = "";
+        } else {
+            clientFilter = filter.substring(0, idx);
+            serverFilter = filter.substring(idx + 1);
+        }
         result = new ArrayList<>();
         for (Server server : servers.values()) {
-            client = client(server.name);
-            result.addAll(Reference.list(client, client.list(filter)));
+            if (server.name.toLowerCase().contains(serverFilter.toLowerCase())) {
+                client = client(server.name);
+                result.addAll(Reference.list(client, client.list(clientFilter)));
+            }
         }
         return result;
     }